[fix] emoji 正则表达式

+++ iToys-kit/core/src/main/kotlin/com/itoys/android/core/CoreConst.kt
@@ -13,7 +13,7 @@ import com.fasterxml.jackson.annotation.JsonProperty
  */
 data class DictModel(
     @JsonProperty("dictKey")
-    val dictKey: Int?,
+    val dictKey: Long?,
     @JsonProperty("dictValue")
     val dictValue: String?,
 ) {
